requirements.txt
flask
numpy
pandas
scikit-learn
imbalanced-learn
joblib
gunicorn
pymongo
dnspython
python-dotenv